python-server/pyproject.toml
[project]
name = "coursewagon-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types>=0.7.0",
    "anyio>=4.9.0",
    "apscheduler>=3.11.0",
    "audioread>=3.0.1",
    "azure-core>=1.35.0",
    "azure-identity>=1.23.1",
    "azure-storage-blob>=12.26.0",
    "bcrypt>=4.3.0",
    "blinker>=1.9.0",
    "cachecontrol>=0.14.3",
    "cachetools>=5.5.2",
    "certifi>=2025.8.3",
    "cffi>=1.17.1",
    "charset-normalizer>=3.4.2",
    "click>=8.2.1",
    "cryptography>=45.0.5",
    "decorator>=5.2.1",
    "dnspython>=2.7.0",
    "ecdsa>=0.19.1",
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "fastapi-jwt-auth>=0.2.0",
    "fastapi-sqlalchemy>=0.2.1",
    "firebase-admin>=7.1.0",
    "google-api-core>=2.25.1",
    "google-api-python-client>=2.177.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-cloud-core>=2.4.3",
    "google-cloud-firestore>=2.21.0",
    "google-cloud-storage>=3.2.0",
    "google-crc32c>=1.7.1",
    "google-genai>=1.28.0",
    "google-resumable-media>=2.7.2",
    "googleapis-common-protos>=1.70.0",
    "greenlet>=3.2.3",
    "grpcio>=1.74.0",
    "grpcio-status>=1.74.0",
    "gunicorn>=23.0.0",
    "h11>=0.16.0",
    "h2>=4.2.0",
    "hpack>=4.1.0",
    "httpcore>=1.0.9",
    "httplib2>=0.22.0",
    "httpx>=0.28.1",
    "hyperframe>=6.1.0",
    "idna>=3.10",
    "isodate>=0.7.2",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "joblib>=1.5.1",
    "lazy-loader>=0.4",
    "librosa>=0.11.0",
    "llvmlite>=0.44.0",
    "markupsafe>=3.0.2",
    "msal>=1.33.0",
    "msal-extensions>=1.3.1",
    "msgpack>=1.1.1",
    "mysql-connector-python>=9.4.0",
    "numba>=0.61.2",
    "numpy>=2.2.6",
    "packaging>=25.0",
    "passlib>=1.7.4",
    "pillow>=11.3.0",
    "platformdirs>=4.3.8",
    "pooch>=1.8.2",
    "proto-plus>=1.26.1",
    "protobuf>=6.31.1",
    "pyasn1>=0.6.1",
    "pyasn1-modules>=0.4.2",
    "pycparser>=2.22",
    "pydantic>=2.11.7",
    "pydantic-core>=2.33.2",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.1",
    "pyparsing>=3.2.3",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "rsa>=4.9.1",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "six>=1.17.0",
    "sniffio>=1.3.1",
    "soundfile>=0.13.1",
    "soxr>=0.5.0.post1",
    "sqlalchemy>=2.0.42",
    "starlette>=0.47.2",
    "tenacity>=8.5.0",
    "threadpoolctl>=3.6.0",
    "tqdm>=4.67.1",
    "typing-extensions>=4.14.1",
    "typing-inspection>=0.4.1",
    "tzlocal>=5.3.1",
    "uritemplate>=4.2.0",
    "urllib3>=2.5.0",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
    "werkzeug>=3.1.3",
]
